Set deleted data true in upload datastore if there is no data to delete or if these data have been removed

+++ src/domain/usecases/data-entry/amc/DeleteAMCProductLevelDataUseCase.ts
@@ -14,6 +14,10 @@ import { getStringFromFileBlob } from "../utils/fileToString";
 import { InstanceRepository } from "../../../repositories/InstanceRepository";
 import { UseCase } from "../../../../CompositionRoot";
 import { AMC_PRODUCT_REGISTER_PROGRAM_ID, AMR_GLASS_AMC_TET_PRODUCT_REGISTER } from "./ImportAMCProductLevelData";
+import { GlassUploadsRepository } from "../../../repositories/GlassUploadsRepository";
+import { GlassUploads } from "../../../entities/GlassUploads";
+import { Id } from "../../../entities/Ref";
+import { AMC_SUBSTANCE_CALCULATED_CONSUMPTION_PROGRAM_ID } from "./ImportAMCSubstanceLevelData";
 
 // NOTICE: code adapted for node environment from ImportAMCProductLevelData.ts (only DELETE)
 export class DeleteAMCProductLevelDataUseCase implements UseCase {
@@ -25,15 +29,12 @@ export class DeleteAMCProductLevelDataUseCase implements UseCase {
             glassDocumentsRepository: GlassDocumentsRepository;
             metadataRepository: MetadataRepository;
             amcSubstanceDataRepository: AMCSubstanceDataRepository;
+            glassUploadsRepository: GlassUploadsRepository;
         }
     ) {}
 
-    public execute(
-        arrayBuffer: ArrayBuffer,
-        eventListId: string | undefined,
-        orgUnitId: string,
-        calculatedEventListFileId?: string
-    ): FutureData<ImportSummary> {
+    public execute(arrayBuffer: ArrayBuffer, upload: GlassUploads): FutureData<ImportSummary> {
+        const { id: uploadId, calculatedEventListFileId, calculatedEventListDataDeleted } = upload;
         return this.options.excelRepository
             .loadTemplateFromArrayBuffer(arrayBuffer, AMC_PRODUCT_REGISTER_PROGRAM_ID)
             .flatMap(_templateId => {
@@ -56,26 +57,29 @@ export class DeleteAMCProductLevelDataUseCase implements UseCase {
                             if (!dataPackage) return Future.error("Cannot find data package");
 
                             return downloadIdsAndDeleteTrackedEntitiesUsingFileBlob(
-                                eventListId,
-                                orgUnitId,
+                                upload,
+                                AMC_PRODUCT_REGISTER_PROGRAM_ID,
                                 "DELETE",
                                 AMR_GLASS_AMC_TET_PRODUCT_REGISTER,
                                 this.options.glassDocumentsRepository,
                                 this.options.trackerRepository,
-                                this.options.metadataRepository
+                                this.options.metadataRepository,
+                                this.options.glassUploadsRepository
                             ).flatMap(deleteProductSummary => {
                                 if (
                                     (deleteProductSummary.status === "SUCCESS" ||
                                         deleteProductSummary.status === "WARNING") &&
-                                    calculatedEventListFileId
+                                    calculatedEventListFileId &&
+                                    !calculatedEventListDataDeleted
                                 ) {
                                     return this.deleteCalculatedSubstanceConsumptionData(
+                                        uploadId,
                                         deleteProductSummary,
                                         calculatedEventListFileId
                                     );
+                                } else {
+                                    return Future.success(deleteProductSummary);
                                 }
-
-                                return Future.success(deleteProductSummary);
                             });
                         });
                     });
@@ -83,48 +87,81 @@ export class DeleteAMCProductLevelDataUseCase implements UseCase {
     }
 
     private deleteCalculatedSubstanceConsumptionData(
+        uploadId: Id,
         deleteProductSummary: ImportSummary,
         calculatedSubstanceConsumptionListFileId: string
-    ) {
+    ): FutureData<ImportSummary> {
         return this.options.glassDocumentsRepository
             .download(calculatedSubstanceConsumptionListFileId)
             .flatMap(eventListFileBlob => {
                 return getStringFromFileBlob(eventListFileBlob).flatMap(_events => {
-                    const calculatedConsumptionIds: string[] = JSON.parse(_events);
-                    return this.options.amcSubstanceDataRepository
-                        .deleteCalculatedSubstanceConsumptionDataById(calculatedConsumptionIds)
-                        .flatMap(deleteCalculatedSubstanceConsumptionResponse => {
-                            return mapToImportSummary(
-                                deleteCalculatedSubstanceConsumptionResponse,
-                                "event",
-                                this.options.metadataRepository
-                            ).flatMap(deleteCalculatedSubstanceConsumptionSummary => {
-                                return Future.success({
-                                    ...deleteCalculatedSubstanceConsumptionSummary.importSummary,
-                                    importCount: {
-                                        imported:
-                                            deleteCalculatedSubstanceConsumptionSummary.importSummary.importCount
-                                                .imported + deleteProductSummary.importCount.imported,
-                                        updated:
-                                            deleteCalculatedSubstanceConsumptionSummary.importSummary.importCount
-                                                .updated + deleteProductSummary.importCount.updated,
-                                        ignored:
-                                            deleteCalculatedSubstanceConsumptionSummary.importSummary.importCount
-                                                .ignored + deleteProductSummary.importCount.ignored,
-                                        deleted:
-                                            deleteCalculatedSubstanceConsumptionSummary.importSummary.importCount
-                                                .deleted + deleteProductSummary.importCount.deleted,
-                                    },
-                                    nonBlockingErrors: [
-                                        ...deleteCalculatedSubstanceConsumptionSummary.importSummary.nonBlockingErrors,
-                                        ...deleteProductSummary.nonBlockingErrors,
-                                    ],
-                                    blockingErrors: [
-                                        ...deleteCalculatedSubstanceConsumptionSummary.importSummary.blockingErrors,
-                                        ...deleteProductSummary.blockingErrors,
-                                    ],
+                    const calculatedConsumptionIds: Id[] = JSON.parse(_events);
+
+                    return this.options.trackerRepository
+                        .getExistingEventsIdsByIds(
+                            calculatedConsumptionIds,
+                            AMC_SUBSTANCE_CALCULATED_CONSUMPTION_PROGRAM_ID
+                        )
+                        .flatMap(existingEventsIds => {
+                            if (!existingEventsIds.length) {
+                                return this.options.glassUploadsRepository
+                                    .setCalculatedEventListDataDeleted(uploadId)
+                                    .flatMap(() => {
+                                        return Future.success(deleteProductSummary);
+                                    });
+                            }
+
+                            return this.options.amcSubstanceDataRepository
+                                .deleteCalculatedSubstanceConsumptionDataById(existingEventsIds)
+                                .flatMap(deleteCalculatedSubstanceConsumptionResponse => {
+                                    return mapToImportSummary(
+                                        deleteCalculatedSubstanceConsumptionResponse,
+                                        "event",
+                                        this.options.metadataRepository
+                                    ).flatMap(deleteCalculatedSubstanceConsumptionSummary => {
+                                        const mergedSummary: ImportSummary = {
+                                            ...deleteCalculatedSubstanceConsumptionSummary.importSummary,
+                                            importCount: {
+                                                imported:
+                                                    deleteCalculatedSubstanceConsumptionSummary.importSummary
+                                                        .importCount.imported +
+                                                    deleteProductSummary.importCount.imported,
+                                                updated:
+                                                    deleteCalculatedSubstanceConsumptionSummary.importSummary
+                                                        .importCount.updated + deleteProductSummary.importCount.updated,
+                                                ignored:
+                                                    deleteCalculatedSubstanceConsumptionSummary.importSummary
+                                                        .importCount.ignored + deleteProductSummary.importCount.ignored,
+                                                deleted:
+                                                    deleteCalculatedSubstanceConsumptionSummary.importSummary
+                                                        .importCount.deleted + deleteProductSummary.importCount.deleted,
+                                            },
+                                            nonBlockingErrors: [
+                                                ...deleteCalculatedSubstanceConsumptionSummary.importSummary
+                                                    .nonBlockingErrors,
+                                                ...deleteProductSummary.nonBlockingErrors,
+                                            ],
+                                            blockingErrors: [
+                                                ...deleteCalculatedSubstanceConsumptionSummary.importSummary
+                                                    .blockingErrors,
+                                                ...deleteProductSummary.blockingErrors,
+                                            ],
+                                        };
+
+                                        if (
+                                            deleteCalculatedSubstanceConsumptionSummary.importSummary.status ===
+                                            "SUCCESS"
+                                        ) {
+                                            return this.options.glassUploadsRepository
+                                                .setCalculatedEventListDataDeleted(uploadId)
+                                                .flatMap(() => {
+                                                    return Future.success(mergedSummary);
+                                                });
+                                        } else {
+                                            return Future.success(mergedSummary);
+                                        }
+                                    });
                                 });
-                            });
                         });
                 });
             });
